KYLIN-4577 Throws UnrecognizedPropertyException when server self discovery enables
Problem:
When server self discovery enables, it throws UnrecognizedPropertyException on some env, for example: cdh 5.7.

Root cause:
The json of ServiceInstance is incompatible, related to the issue of curator:  https://issues.apache.org/jira/browse/CURATOR-394
